Fix typo and rename deploymentmetadata
