+++ controllers/oneagent/daemonset.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	dynatracev1alpha1 "github.com/Dynatrace/dynatrace-operator/api/v1alpha1"
-	"github.com/Dynatrace/dynatrace-operator/deployment_metadata"
+	"github.com/Dynatrace/dynatrace-operator/deploymentmetadata"
 	"github.com/Dynatrace/dynatrace-operator/version"
 )
 
@@ -24,7 +24,7 @@ func prepareArgs(instance *dynatracev1alpha1.DynaKube, fs *dynatracev1alpha1.Ful
 
 	args = append(args, "--set-host-property=OperatorVersion="+version.Version)
 
-	metadata := deployment_metadata.NewDeploymentMetadata(
+	metadata := deploymentmetadata.NewDeploymentMetadata(
 		version.Version,
 		clusterID,
 		instance.Status.OneAgent.Version,
